refactor: change requests
